3.5.0 release (#40)
* implement logging of latent genotypes (#14)

* Fix NA values in allele frequency vector (#20)

The code now checks for NA values in the allele frequency vector and replaces them with 0. A warning message is displayed if NA values are detected, indicating a potential problem with the MCMC chain or loci with no diversity.

* Add start_profiler and stop_profiler functions (#24)

- Added start_profiler function to R/RcppExports.R
- Added stop_profiler function to R/RcppExports.R
- Modified src/Makevars to include ENABLE_PROFILER flag and link with -lprofiler
- Added start_profiler and stop_profiler functions to src/RcppExports.cpp
- Created new file src/profiler.cpp containing the implementation of start_profiler and stop_profiler functions

* Refactor Chain class to include methods for calculating likelihood, prior, and posterior for a specific sample. (#33)

- Add `get_llik(int sample)` method to calculate the likelihood for a specific sample.
- Add `get_prior(int sample)` method to calculate the prior for a specific sample.
- Add `get_posterior(int sample)` method to calculate the posterior for a specific sample.

These changes allow for more granular calculations of likelihood, prior, and posterior values based on individual samples.

* Implement maximum runtime (#34)

* Implement maximum runtime

* log chain runtime

* add utility to detect if openmp is enabled

* ungroup input data when loading long form (#37)

* fix memory crash in prob_any_missing (#38)

The code fix addresses a memory crash issue in the `prob_any_missing` function. The fix includes adding a condition to check if `maxNumEvents` is less than `totalEvents`. If true, it fills the `probVec` vector with 1.0 and returns it. This prevents the memory crash and ensures proper execution of the function.

* restore accidentally removed openmp library to compilation

* Add new datasets namibia_data and regional_allele_frequencies. Update vignette with code to use the new datasets. (#27)

Add new datasets namibia_data and regional_allele_frequencies. Update vignette with code to use the new datasets.

- Added new datasets namibia_data and regional_allele_frequencies.
- Updated vignette with code to use the new datasets.

* update citation and manuscript